ingestor-service/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elastic-resume-base-ingestor"
version = "1.0.0"
description = "Resume ingestor microservice for Elastic Resume Base"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.3.4",
    "google-auth==2.29.0",
    "google-api-python-client==2.128.0",
    "google-cloud-firestore==2.16.1",
    "google-cloud-pubsub==2.21.4",
    "pypdf==4.2.0",
    "elastic-resume-base-hermes",
    "elastic-resume-base-bowltie",
    "elastic-resume-base-toolbox",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ---------------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------------

[tool.black]
line-length = 100
target-version = ["py311"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
